scripts/requirements.txt
influxdb==5.3.1
pandas==2.1.3
python-dotenv==1.0.0